Add {{ link(to="PageTitle") }} shortcode
This makes it easier to make links to other pages, as it now works in a
way that more closely represents Wiki-style.

To make this work, a script is added that creates a linkmap.toml file.
This needs to be done ahead of time due to how Zola's shortcode system
works, so we use the data import feature of Zola.